Изменен тип, чтобы не было лишнего приведения типов.
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@62383 954022d7-b5bf-4e40-9824-e11837661b57

+++ DesktopEditor/common/File.h
@@ -794,7 +794,7 @@ namespace NSFile
 
 			std::wstring sRet(pBuffer);
 
-			int nSeparatorPos = sRet.find_last_of(wchar_t('/'));
+			size_t nSeparatorPos = sRet.find_last_of(wchar_t('/'));
 			if (std::wstring::npos == nSeparatorPos)
 			{
 				nSeparatorPos = sRet.find_last_of(wchar_t('\\'));
